chatbot_train/requirements.txt
certifi==2018.1.18
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.6
filelock==3.4.1
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.8.3
joblib==1.1.0
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==0.24.2
scipy==1.9.1
semantic-version==2.9.0
setuptools-rust==1.1.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tokenizers==0.11.5
torch==1.11.0
torchinfo==1.5.4
tqdm==4.62.3
transformers==4.16.2
typing_extensions==4.1.1
urllib3==1.26.8
zipp==3.6.0
